% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcf.R
\name{addcf}
\alias{addcf}
\title{Calculate Additive Change Factor}
\usage{
addcf(baselinedata, futuredata)
}
\arguments{
\item{baselinedata}{data.frame with the day of year and daily average value for the baseline period}

\item{futuredata}{data.frame with the day of year and daily average value for the future period}
}
\value{
data.frame of additive change factor for each day of the year
}
\description{
Calculate Additive Change Factor
}

