% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyaddcf.R
\name{applyaddcf}
\alias{applyaddcf}
\title{Apply Additive Change Factor}
\usage{
applyaddcf(obsdata, addcfdata, period, yearsdiff)
}
\arguments{
\item{obsdata}{data.frame with the date and observed precipitation record}

\item{addcfdata}{data.frame with day of the year and change factor}

\item{period}{vector of start and end years for historical observed period (should be same as baseline period)}

\item{yearsdiff}{number of years between baseline and future periods}
}
\value{
data.frame of scaled projection of precipitation
}
\description{
Apply Additive Change Factor
}

