% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applymultcf.R
\name{applymultcf}
\alias{applymultcf}
\title{Apply Multiplicative Change Factor}
\usage{
applymultcf(obsdata, multcfdata, period, yearsdiff)
}
\arguments{
\item{obsdata}{data.frame with the date and observed precipitation record}

\item{multcfdata}{data.frame with day of the year and change factor}

\item{period}{vector of start and end years for historical observed period (should be same as baseline period)}

\item{yearsdiff}{number of years between baseline and future periods}
}
\value{
data.frame of scaled projection of precipitation
}
\description{
Apply Multiplicative Change Factor
}

