% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgdaily.R
\name{avgdaily}
\alias{avgdaily}
\title{Calculate Precipitation by Day of the Year}
\usage{
avgdaily(data)
}
\arguments{
\item{data}{data.frame with the date and daily precipitation projections}
}
\value{
data.frame Data frame of averaged precipitation values
}
\description{
Calculate Precipitation by Day of the Year
}

