% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calccf.R
\name{calccf}
\alias{calccf}
\title{Calculate Change Factor}
\usage{
calccf(baselinedata, futuredata, cftype)
}
\arguments{
\item{baselinedata}{data.frame with the day of year and daily average value for the baseline period}

\item{futuredata}{data.frame with the day of year and daily average value for the future period}

\item{type}{of change factor to be calculated}
}
\value{
data.frame of additive or multiplicative change factor for each day of the year
}
\description{
Calculate Change Factor
}

