% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcecdf.R
\name{calcecdf}
\alias{calcecdf}
\title{Calculate ECDF of precipitation values}
\usage{
calcecdf(data, precipcol)
}
\arguments{
\item{precipcol}{string representing the column in data frame with precipitation values}

\item{data.frame}{containing precipitation values}
}
\value{
data.frame Data frame with percentiles (exact and rounded)
}
\description{
Calculate ECDF of precipitation values
}

