% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfm.R
\name{ccfm}
\alias{ccfm}
\title{Apply CCFM}
\usage{
ccfm(data, precipcol, addcfcol, multcfcol, upperboundadd, threshold)
}
\arguments{
\item{data, }{data.frame}

\item{precipcol}{string representing the column in data frame with precipitation values}

\item{addcfcol}{string, name of column with additive change factors}

\item{multcfcol}{string, name of column with multiplicative change factors}

\item{upperboundadd}{numeric value (0-100) which is the upper bound percentile with which to apply the additive change factor}

\item{threshold}{numeric value (0-100) above which the additive change factor is applied}
}
\value{
data Data frame with percentiles (exact and rounded)
}
\description{
Apply CCFM
}

