% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatperiods.R
\name{formatperiods}
\alias{formatperiods}
\title{Formats baseline and historic projections for a chosen model}
\usage{
formatperiods(data, modelname, period)
}
\arguments{
\item{data}{data.frame with formatted time series of projections for various models}

\item{modelname}{Model and RCP}

\item{period}{vector, Bounding years for the baseline or future period}
}
\value{
data.frame Data frame of projections for the specified period
}
\description{
Formats baseline and historic projections for a chosen model
}

