% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listmodels.R
\name{listmodels}
\alias{listmodels}
\title{Shows available models for a particular dataset}
\usage{
listmodels(data)
}
\arguments{
\item{data}{data.frame with formatted time series of projections for various models}
}
\value{
list list of model names
}
\description{
Shows available models for a particular dataset
}

