% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multcf.R
\name{multcf}
\alias{multcf}
\title{Calculate Multiplicative Change Factor}
\usage{
multcf(baselinedata, futuredata)
}
\arguments{
\item{baselinedata}{data.frame with the day of year and daily average value for the baseline period}

\item{futuredata}{data.frame with the day of year and daily average value for the future period}
}
\value{
data.frame of multiplicative change factor for each day of the year
}
\description{
Calculate Multiplicative Change Factor
}

