% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipsummary.R
\name{precipsummary}
\alias{precipsummary}
\title{PrecipSummary}
\usage{
precipsummary(data, precipcol)
}
\arguments{
\item{data, }{data.frame}

\item{precipcol}{string representing the column in data frame with precipitation (observed, modeled, or scaled) values}
}
\description{
PrecipSummary
}

