% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readobs.R
\name{read.obs}
\alias{read.obs}
\title{Read Observed Data}
\usage{
read.obs(observed, city)
}
\arguments{
\item{observed, }{data.frame}
}
\value{
obsdata Data frame with observations for specified city
}
\description{
Read Observed Data
}

