% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readobs.R
\name{readobs}
\alias{readobs}
\title{Read Observed Data}
\usage{
readobs(observed, city)
}
\arguments{
\item{observed, }{data.frame with historical observed data}

\item{city, }{location (currently limited to SLC,Toledo, Phoenix, Seattle, Miami, and Houston)}
}
\value{
obsdata, Data frame with observations for specified city
}
\description{
Read Observed Data
}

