% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalemultiple.R
\name{scalemultiple}
\alias{scalemultiple}
\title{Applies multiple change factors to observed precipitation}
\usage{
scalemultiple(observed, cftype, percent)
}
\arguments{
\item{observed, }{data.frame with observed time series (formatted to include day of the year)}

\item{cftype, }{string indicating 'additive' or 'multiplicative'}

\item{percent, }{data.frame with change factors by percentile}
}
\value{
data.frame Data frame of scaled projections
}
\description{
Applies multiple change factors to observed precipitation
}

