% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalesingle.R
\name{scalesingle}
\alias{scalesingle}
\title{Applies single change factors to observed precipitation}
\usage{
scalesingle(cf, observed, cftype)
}
\arguments{
\item{cf}{data.frame with calculated change factors by day of the year}

\item{observed, }{data.frame with observed time series (formatted to include day of the year)}

\item{cftype, }{string indicating 'additive' or 'multiplicative'}
}
\value{
data.frame Data frame of scaled projections
}
\description{
Applies single change factors to observed precipitation
}

