% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatperiods.R
\name{formatperiods}
\alias{formatperiods}
\title{Formats baseline and historic projections for a chosen model}
\usage{
formatperiods(data, modelname, baseline, future)
}
\arguments{
\item{data}{data.frame with formatted time series of projections for various models}

\item{modelname}{Model and RCP}

\item{baseline}{vector, Bounding years for the baseline period, should be equal in length to the future period}

\item{future}{vector, Bounding years for the future period}
}
\value{
list List of projections for baseline and future periods
}
\description{
Formats baseline and historic projections for a chosen model
}

